Support new account-gql
